refactor(core): dir structure
refactor directory structure based on https://github.com/golang-standards/project-layout
