+++ README.md
@@ -61,8 +61,6 @@
 <!-- ABOUT THE PROJECT -->
 ## About The Project
 
-[![Product Name Screen Shot][product-screenshot]](https://example.com)
-
 This package provides a unified command line interface to Arvan CDN Services.
